# perf: more efficient treemap implementation for row ids