--- a/rust/lance/src/io/exec/knn.rs
+++ b/rust/lance/src/io/exec/knn.rs
@@ -28,10 +28,9 @@ use datafusion::physical_plan::{
 };
 use futures::stream::Stream;
 use futures::{FutureExt, StreamExt, TryStreamExt};
-use lance_core::utils::mask::RowIdMask;
+use lance_core::utils::mask::{RowIdMask, RowIdTreeMap};
 use lance_core::{ROW_ID, ROW_ID_FIELD};
 use lance_index::vector::{flat::flat_search, Query, DIST_COL};
-use roaring::RoaringTreemap;
 use snafu::{location, Location};
 use tokio::sync::mpsc::Receiver;
 use tokio::task::JoinHandle;
@@ -257,7 +256,7 @@ struct FilteredRowIdsToPrefilter(SendableRecordBatchStream);
 #[async_trait]
 impl FilterLoader for FilteredRowIdsToPrefilter {
     async fn load(mut self: Box<Self>) -> Result<RowIdMask> {
-        let mut allow_list = RoaringTreemap::new();
+        let mut allow_list = RowIdTreeMap::new();
         while let Some(batch) = self.0.next().await {
             let batch = batch?;
             let row_ids = batch.column_by_name(ROW_ID).expect(