Provides our own TreeMap implementation, which is optimized for the case where an entire fragment is added. When a full fragment is to be marked, we use `None` instead of a full `RoaringBitmap`.

Fixes #1633