@wjones127 2023-11-18
Running the repro described in #1633, I get:

**Before:**
```
creating table
appending
appending
creating index
testing query
3 query took 0.00442233300418593 seconds
delete a fragement
testing query
3 query took 0.44909045900567435 seconds
```

**After:**
```
creating table
appending
appending
creating index
testing query
3 query took 0.015102374978596345 seconds
delete a fragement
testing query
3 query took 0.0017974170041270554 seconds
```