@wjones127 2023-11-18
The benchmarks I ran were on different tables for each of them. I've refactored so I can run on the same table and get more samples.

Here are results

**Before:**

```
testing query before deletion
avg query took 0.0004714387501007877 seconds
testing query
avg query took 0.1331728041599854 seconds
```

**After:**

```
testing query before deletion
avg query took 0.0004256525001255795 seconds
testing query
avg query took 0.0004011687499587424 seconds
```


<details><summary>datagen.py</summary>
<p>

```python
import lance
import pyarrow as pa

import uuid
import random

def schema():
    return pa.schema([
        pa.field('id', pa.string()),
        pa.field('vector', pa.list_(pa.float32(), 128)),
    ])

def random_row():
    return {
        'id': uuid.uuid4().hex,
        'vector': [random.random() for _ in range(128)],
    }

def random_table(n):
    return pa.Table.from_pylist([random_row() for _ in range(n)], schema=schema())

print("creating table")
lance.write_dataset(random_table(0), "test.lance")

print("appending")
lance.write_dataset(random_table(512), "test.lance", mode="append")

print("appending")
second_append = random_table(2)
lance.write_dataset(second_append, "test.lance", mode="append")

print("creating index")
ds = lance.dataset("test.lance")
ds.create_index("vector", "IVF_PQ", metric="L2", num_partitions=1, num_sub_vectors=16)

print("delete a fragement")
ids_to_delete = second_append["id"].to_pylist()
quoted_ids = ", ".join([f'"{id}"' for id in ids_to_delete])
ds = lance.dataset("test.lance")
ds.delete(f"`id` IN ({quoted_ids})")
```

</p>
</details> 

<details><summary>repro.py</summary>
<p>


```python
import lance

import random
import time


def test_query(before_deletion=True):
    ds = lance.dataset("test.lance")
    if before_deletion:
        ds = lance.dataset("test.lance", version=ds.version - 1)
    start = time.perf_counter()
    for _ in range(100):
        query = [random.random() for _ in range(128)]
        ds.to_table(nearest={
            "column": "vector",
            "q": query,
            "use_index": True
        })
    total_time = time.perf_counter() - start
    avg_time = total_time / 100
    print(f"avg query took {avg_time} seconds")


print("testing query before deletion")
test_query(before_deletion=True)

print("testing query")
test_query(before_deletion=False)
```
</p>
</details>